created commitInfosParser
