+++ src/test/java/com/urswolfer/gerrit/client/rest/http/changes/ChangesRestClientTest.java
@@ -88,8 +88,7 @@ public class ChangesRestClientTest {
 
         ChangesRestClient changes = new ChangesRestClient(gerritRestClient, changesParser, null, null,
             null, null, null, null,
-            null, null, null, null,
-            null, null, null, null);
+            null, null, null, null, null);
 
         Changes.QueryRequest queryRequest = changes.query();
         testCase.queryParameter.apply(queryRequest).get();
@@ -107,8 +106,7 @@ public class ChangesRestClientTest {
 
         ChangesRestClient changesRestClient = new ChangesRestClient(gerritRestClient, changesParser, commentsParser, null,
             null, null, null, null, null,
-            null, null, null, null,
-            null, null, null);
+            null, null, null, null);
         changesRestClient.query("is:open").get();
 
         EasyMock.verify(gerritRestClient);
@@ -122,8 +120,7 @@ public class ChangesRestClientTest {
 
         ChangesRestClient changesRestClient = new ChangesRestClient(gerritRestClient, changesParser, commentsParser, null,
             null, null, null, null, null,
-            null, null, null, null,
-            null, null, null);
+            null, null, null, null);
 
         ChangeApi changeApi = changesRestClient.id(123);
 
@@ -138,8 +135,7 @@ public class ChangesRestClientTest {
 
         ChangesRestClient changesRestClient = new ChangesRestClient(gerritRestClient, changesParser, commentsParser, null,
             null, null, null, null, null,
-            null, null, null, null,
-            null, null, null);
+            null, null, null, null);
 
         ChangeApi changeApi = changesRestClient.id("packages/test", 123);
 
@@ -154,8 +150,7 @@ public class ChangesRestClientTest {
 
         ChangesRestClient changesRestClient = new ChangesRestClient(gerritRestClient, changesParser, commentsParser, null,
             null,  null, null, null, null,
-            null, null, null, null,
-            null, null, null);
+            null, null, null, null);
 
         ChangeApi changeApi = changesRestClient.id("packages/test", "master", "Ieabd72e73f3da0df90fd6e8cba8f6c5dd7d120df");
 
@@ -170,8 +165,7 @@ public class ChangesRestClientTest {
 
         ChangesRestClient changes = new ChangesRestClient(gerritRestClient, changesParser,null, null,
             null, null, null, null, null,
-            null, null, null, null,
-            null, null, null);
+            null, null, null, null);
 
         changes.query().get();
 
@@ -192,8 +186,7 @@ public class ChangesRestClientTest {
 
         ChangesRestClient changes = new ChangesRestClient(gerritRestClient, changesParser, null, null,
             null, null, null, null, null,
-            null, null, null, null,
-            null, null, null);
+            null, null, null, null);
 
         ChangeApi changeApi = changes.create(changeInput);
