+++ src/test/java/com/urswolfer/gerrit/client/rest/http/changes/IncludedInCommitInfosParserTest.java
@@ -23,7 +23,7 @@ import com.google.gerrit.extensions.api.changes.IncludedInInfo;
 import com.google.gson.JsonElement;
 import com.urswolfer.gerrit.client.rest.http.common.AbstractParserTest;
 
-public class IncludedInInfoParserTest extends AbstractParserTest {
+public class IncludedInCommitInfosParserTest extends AbstractParserTest {
     private final IncludedInInfoParser includedInInfoParser = new IncludedInInfoParser(getGson());
 
     @Test
